Extended state machine and added background color
The state machine now supports expicit starting and ending of presentation mode.  The previous slide function now also starts a slide from the top when it contains pauses and it goes to the previous slide when a slide does not have pauses.
The background color can set (currently with border the same as background).